pyproject.toml
[tool.poetry]
name = "babble"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
ecdsa = "^0.18.0"
requests = "^2.28.2"
bech32 = "^1.2.0"
pycryptodome = "^3.16.0"
pyjwt = "^2.6.0"
eciespy = "^0.3.13"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
